move auth to top level
