+++ src/components/Chat/ChatInterface.tsx
@@ -1,6 +1,6 @@
 import "./index.css";
 import React from "react";
-import { Paper, Button } from "@mui/material";
+import { Paper } from "@mui/material";
 import { TextInput } from "./TextInput";
 import { MessageOther, MessageSelf } from "./Message";
 import { useState } from "react";
@@ -24,11 +24,7 @@ const initMsgs = [
     />
 ]
 
-type Props = {
-    logout: () => void
-}
-
-export default function ChatInterface({logout}: Props) {
+export default function ChatInterface() {
 
     const [rawMsgs, setRawMsgs] = useState(initRawMsgs);
 
@@ -120,15 +116,6 @@ export default function ChatInterface({logout}: Props) {
     }
     return (
         <div className="chatContainer">
-            <div className="logoutBtnContainer">
-                <Button
-                    className="logoutBtn"
-                    sx={{ bgcolor: "var(--ifm-color-primary)", color: "white" }}
-                    onClick={() => logout()}
-                >
-                    End Tutoring Session
-                </Button>
-            </div>
             <Paper className="paper" elevation={0}>
                 <Paper id="style-1" className="messagesBody">
                     {showSpinner ? spinner : <></>}
